Revent credits scene. Fix sounds bug
