--- Assets/ButtonSounds.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 522b0c1665b4fd344a4038b7459b3199
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
