+++ Assets/Scripts/ButtonSoundsScript.cs
@@ -1,18 +1,29 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
-public class ButtonSounds : MonoBehaviour
+public class ButtonSoundsScript : MonoBehaviour
 {
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
+	public AudioSource clickSoundFX;
+	public AudioClip hoverSound;
+	public AudioClip clickSound;
+	
+	public void HoverSound()
+	{
+		clickSoundFX.PlayOneShot(hoverSound);
+	}
 
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
+	public void ClickSound()
+	{
+		clickSoundFX.PlayOneShot(clickSound);
+	}
+
+	private void Update()
+	{
+		if(clickSoundFX.isActiveAndEnabled != true)
+		{
+			clickSoundFX = GameObject.Find("AudioManager").GetComponent<AudioSource>();
+		}
+	}
 }
